キーイベントやホイールイベントを全体で受け取るようにする
