+++ src/ButtonArea.tsx
@@ -3,7 +3,7 @@ import { Fab, Tooltip } from '@mui/material';
 import React from 'react';
 import { useRecoilCallback, useSetRecoilState } from 'recoil';
 import { Add, Description, HorizontalRule } from '@mui/icons-material';
-import { edgeListAtom, modeSelector, nodeListAtom } from './atoms';
+import { edgeListAtom, modeAtom, modeSelector, nodeListAtom } from './atoms';
 import { Mode } from './helpers/modehelper';
 
 const ButtonArea: React.FC = () => {
@@ -13,6 +13,8 @@ const ButtonArea: React.FC = () => {
       () => {
         const nodeList = snapshot.getLoadable(nodeListAtom).getValue();
         const edgeList = snapshot.getLoadable(edgeListAtom).getValue();
+        const mode = snapshot.getLoadable(modeAtom).getValue();
+        console.log('mode: ', mode);
         console.log('nodeList: ', nodeList);
         console.log('edgeList: ', edgeList);
       },
