+++ src/DrawArea.tsx
@@ -1,17 +1,16 @@
 /* eslint-disable jsx-a11y/tabindex-no-positive */
 /* eslint-disable jsx-a11y/no-noninteractive-tabindex */
 /* eslint-disable jsx-a11y/no-static-element-interactions */
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect } from 'react';
 import { Stage, Layer } from 'react-konva';
 import './App.css';
 import { useRecoilBridgeAcrossReactRoots_UNSTABLE, useRecoilState } from 'recoil';
-import { nextType } from './symbols';
 import Grid from './Grid';
 import { Symbol } from './Symbol';
-import { add, RealPoint, sub, toFixedVirtualGrid, toVirtualGrid } from './helpers/gridhelper';
+import { toFixedVirtualGrid } from './helpers/gridhelper';
 import { useWindowSize } from './hooks/useWindowSize';
 import { useWire } from './hooks/useWire';
-import { Mode, modeToCursorStyle } from './helpers/modehelper';
+import { Mode } from './helpers/modehelper';
 import { usePreviousMode } from './hooks/usePreviousMode';
 import { symbolsAtom, modeAtom, pitchAtom, symbolTypeAtom, upperLeftAtom } from './atoms';
 import Wire from './Wire';
@@ -22,12 +21,12 @@ const DrawArea: React.FC = () => {
   const { height, width } = useWindowSize();
 
   const { setPoint, setPreview, resetSelect } = useWire();
-  const [pitch, setPitch] = useRecoilState(pitchAtom);
-  const [upperLeft, setUpperLeft] = useRecoilState(upperLeftAtom);
+  const [pitch] = useRecoilState(pitchAtom);
+  const [upperLeft] = useRecoilState(upperLeftAtom);
   const [symbols, setSymbols] = useRecoilState(symbolsAtom);
-  const [mode, setMode] = useRecoilState(modeAtom);
+  const [mode] = useRecoilState(modeAtom);
   const prevMode = usePreviousMode(mode);
-  const [symbolType, setSymbolType] = useRecoilState(symbolTypeAtom);
+  const [symbolType] = useRecoilState(symbolTypeAtom);
 
   // mode unmount
   useEffect(() => {
@@ -54,119 +53,58 @@ const DrawArea: React.FC = () => {
   }, [mode]);
   // mode mount
 
-  const divref = useRef<HTMLDivElement>(null);
-
-  const onWheel = (e: WheelEvent) => {
-    e.preventDefault();
-
-    if (e.ctrlKey) {
-      const pos: RealPoint = { x: e.clientX, y: e.clientY };
-      const rawVPos = toVirtualGrid(pos, pitch, upperLeft);
-
-      let newPitch = pitch;
-      if (e.deltaY < 0) newPitch += 1; // Zoom in
-      else newPitch -= 1; // Zoom out
-
-      newPitch = Math.min(Math.max(5, newPitch), 50);
-      const newRawVPos = toVirtualGrid(pos, newPitch, upperLeft);
-
-      setPitch(newPitch);
-      setUpperLeft(add(upperLeft, sub(rawVPos, newRawVPos)));
-    } else {
-      setUpperLeft(add(upperLeft, toVirtualGrid({ x: e.deltaX, y: e.deltaY }, pitch, { vx: 0, vy: 0 })));
-    }
-  };
-
-  useEffect(() => {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
-    divref.current?.addEventListener('wheel', onWheel, { passive: false });
-    return () => {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
-      divref.current?.removeEventListener('wheel', onWheel);
-    };
-  });
-
   return (
-    <div
-      ref={divref}
-      tabIndex={1}
-      onKeyDown={(e) => {
-        switch (e.code) {
-          case 'Escape':
-            setMode(Mode.NONE);
+    <Stage
+      width={width}
+      height={height}
+      onClick={(e) => {
+        const stage = e.target.getStage();
+        if (!stage) return;
+        const pos = stage.getPointerPosition();
+        if (!pos) return;
+
+        const vpos = toFixedVirtualGrid(pos, pitch, upperLeft);
+        switch (mode) {
+          case Mode.WIRE:
+            setPoint(vpos);
             break;
-          case 'KeyL':
-            setMode(Mode.WIRE);
-            break;
-          case 'KeyP':
-            if (mode === Mode.SYMBOL) setSymbolType(nextType(symbolType));
-            else setMode(Mode.SYMBOL);
-            break;
-          case 'KeyE':
-            setPitch(pitch + 1);
-            break;
-          case 'KeyR':
-            setPitch(pitch - 1);
+          case Mode.SYMBOL:
+            setSymbols(
+              symbols
+                .slice(0, -1)
+                .concat({ type: symbolType, point: vpos, key: `symbol_${symbols.length}` })
+                .concat({ type: symbolType, point: vpos, key: `symbol_${symbols.length + 1}` })
+            );
             break;
           default:
         }
       }}
-      style={{ cursor: modeToCursorStyle(mode) }}
+      onMouseMove={(e) => {
+        const stage = e.target.getStage();
+        if (!stage) return;
+        const pos = stage.getPointerPosition();
+        if (!pos) return;
+
+        const vpos = toFixedVirtualGrid(pos, pitch, upperLeft);
+        switch (mode) {
+          case Mode.WIRE:
+            setPreview(vpos);
+            break;
+          case Mode.SYMBOL:
+            setSymbols(symbols.slice(0, -1).concat({ type: symbolType, point: vpos, key: `symbol_${symbols.length}` }));
+            break;
+          default:
+        }
+      }}
     >
-      <Stage
-        width={width}
-        height={height}
-        onClick={(e) => {
-          const stage = e.target.getStage();
-          if (!stage) return;
-          const pos = stage.getPointerPosition();
-          if (!pos) return;
-
-          const vpos = toFixedVirtualGrid(pos, pitch, upperLeft);
-          switch (mode) {
-            case Mode.WIRE:
-              setPoint(vpos);
-              break;
-            case Mode.SYMBOL:
-              setSymbols(
-                symbols
-                  .slice(0, -1)
-                  .concat({ type: symbolType, point: vpos, key: `symbol_${symbols.length}` })
-                  .concat({ type: symbolType, point: vpos, key: `symbol_${symbols.length + 1}` })
-              );
-              break;
-            default:
-          }
-        }}
-        onMouseMove={(e) => {
-          const stage = e.target.getStage();
-          if (!stage) return;
-          const pos = stage.getPointerPosition();
-          if (!pos) return;
-
-          const vpos = toFixedVirtualGrid(pos, pitch, upperLeft);
-          switch (mode) {
-            case Mode.WIRE:
-              setPreview(vpos);
-              break;
-            case Mode.SYMBOL:
-              setSymbols(
-                symbols.slice(0, -1).concat({ type: symbolType, point: vpos, key: `symbol_${symbols.length}` })
-              );
-              break;
-            default:
-          }
-        }}
-      >
-        <Bridge>
-          <Layer>
-            <Grid />
-            <Wire />
-            <Symbol />
-          </Layer>
-        </Bridge>
-      </Stage>
-    </div>
+      <Bridge>
+        <Layer>
+          <Grid />
+          <Wire />
+          <Symbol />
+        </Layer>
+      </Bridge>
+    </Stage>
   );
 };
